nip-10: fix duplicate replies in feed
some replies rendered twice in different positions, seems to be
related to deprecated positional event tags and a regression
introduced in a5961218218ab956b13a2437c6f16c658a88b610

on receiving events it analizes the event tags and stores the id
of the replied event so the client can easily search for replies
later. marked tags are prefered with a fallback to positional tags
as described in nip-10.

mentions are ignored at the moment.

example event that had some replies rendered twice:
22e4ea80161ac591059da611d3ab63c583cb1d47a706826db2fc6955ac0a70b5